<chore> 최종 커밋 - 2 : 품목 정보 등록 가능한 코드 버그 존재

+++ src/main/java/com/example/tae/repository/DummyRepository/PartRepository.java
@@ -3,5 +3,5 @@ package com.example.tae.repository.DummyRepository;
 import com.example.tae.entity.DummyData.Classification.Part;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PartRepository extends JpaRepository<Part, Integer> {
+public interface PartRepository extends JpaRepository<Part, Long> {
 }
